update import score
